+++ tests/common-stub.sh
@@ -15,6 +15,7 @@ EXIT="return"
 # test-pot-nosnap			no		multi	1		inherit		undef	inherit
 # test-pot-single			no		single	0		10.1.2.3	yes		public-bridge
 # test-pot-single-run		yes		single	0		undef		yes		public-bridge
+# test-pot-single-2			no		single	0		10.1.2.3	yes		public-bridge
 # test-pot-3				no		multi	1		10.1.2.3	no		alias
 
 # test-pot-multi-inherit	no		multi	1					no		inherit
@@ -69,7 +70,7 @@ _is_pot()
 		test-pot|test-pot-run|\
 		test-pot-2|test-pot-run-2|\
 		test-pot-0|test-pot-nosnap|test-pot-3|\
-		test-pot-single|test-pot-single-run|\
+		test-pot-single|test-pot-single-run|test-pot-single-2|\
 		test-pot-multi-inherit|test-pot-multi-private|\
 		${POT_DNS_NAME})
 			return 0 # true
@@ -136,7 +137,7 @@ _get_conf_var()
 		test-pot-2|test-pot-run-2)
 			echo "2"
 			;;
-		test-pot-0|test-pot-single|test-pot-single-run)
+		test-pot-0|test-pot-single|test-pot-single-run|test-pot-single-2)
 			echo "0"
 			;;
 		esac
@@ -157,7 +158,7 @@ _get_conf_var()
 			echo ""
 			;;
 		test-pot-2|test-pot-3|\
-		test-pot-single)
+		test-pot-single|test-pot-single-2)
 			echo "10.1.2.3"
 			;;
 		test-pot-multi-private)
@@ -175,7 +176,7 @@ _get_conf_var()
 			echo "alias"
 			;;
 		test-pot-2|test-pot-single-run|\
-		test-pot-single)
+		test-pot-single|test-pot-single-2)
 			echo "public-bridge"
 			;;
 		test-pot-multi-private)
@@ -191,14 +192,14 @@ _get_conf_var()
 		test-pot-multi-inherit|test-pot-multi-private)
 			echo "multi"
 			;;
-		test-pot-single|test-pot-single-run)
+		test-pot-single|test-pot-single-run|test-pot-single-2)
 			echo "single"
 		esac
 		;;
 	"vnet")
 		case $1 in
 		test-pot-2|test-pot-single|test-pot-single-run|\
-		test-pot-multi-private)
+		test-pot-multi-private|test-pot-single-2)
 			echo "true"
 			;;
 		test-pot-3|test-pot-multi-inherit)